+++ app/Http/Controllers/People/Bank/CouponTypesController.php
@@ -60,6 +60,7 @@ class CouponTypesController extends Controller
         $coupon->enabled = isset($request->enabled);
         $coupon->returnable = isset($request->returnable);
         $coupon->qr_code_enabled = isset($request->qr_code_enabled);
+        $coupon->allow_for_helpers = isset($request->allow_for_helpers);
         $coupon->save();
 
         return redirect()->route('coupons.index')
@@ -118,6 +119,7 @@ class CouponTypesController extends Controller
         $coupon->enabled = isset($request->enabled);
         $coupon->returnable = isset($request->returnable);
         $coupon->qr_code_enabled = isset($request->qr_code_enabled);
+        $coupon->allow_for_helpers = isset($request->allow_for_helpers);
         $coupon->save();
 
         return redirect()->route('coupons.show', $coupon)