+++ resources/views/bank/coupons/edit.blade.php
@@ -47,6 +47,7 @@
                 {{ Form::bsCheckbox('enabled', 1, null, __('app.enabled')) }}
                 {{ Form::bsCheckbox('returnable', 1, null, __('people.returnable')) }}
                 {{ Form::bsCheckbox('qr_code_enabled', 1, null, __('people.with_qr_code')) }}
+                {{ Form::bsCheckbox('allow_for_helpers', 1, null, __('people.allow_for_helpers')) }}
             </div>
         </div>