+++ resources/views/bank/coupons/index.blade.php
@@ -19,6 +19,7 @@
                         <th>@lang('app.enabled')</th>
                         <th>@lang('people.returnable')</th>
                         <th>@lang('people.qr_code')</th>
+                        <th>@lang('people.helpers')</th>
                     </tr>
                 </thead>
                 <tbody>
@@ -37,6 +38,7 @@
                             <td>@if($coupon->enabled) @icon(check) @else @icon(times) @endif</td>
                             <td>@if($coupon->returnable) @icon(check) @else @icon(times) @endif</td>
                             <td>@if($coupon->qr_code_enabled) @icon(check) @else @icon(times) @endif</td>
+                            <td>@if($coupon->allow_for_helpers) @icon(check) @else @icon(times) @endif</td>
                         </tr>
                     @endforeach
                 </tbody>