+++ resources/views/bank/coupons/show.blade.php
@@ -75,6 +75,12 @@
         </li>
         <li class="list-group-item">
             <div class="row">
+                <div class="col-sm"><strong>@lang('people.allow_for_helpers')</strong></div>
+                <div class="col-sm">@if($coupon->allow_for_helpers) @icon(check) @else @icon(times) @endif</div>
+            </div>
+        </li>
+        <li class="list-group-item">
+            <div class="row">
                 <div class="col-sm"><strong>@lang('app.registered')</strong></div>
                 <div class="col-sm">{{ $coupon->created_at }}</div>
             </div>